feat(dashboard): all forms and controller works...
- dashboard global view display a table for each entity, with essential
columns, and buttons to `modify` or `delete` for each line, and a global
button `new` for that entity.
- the `delete` button goes to a controller who delete the line from
database.
- the `modify` button goes to a form with all fields of that entity,
corresponding to the columns of the table in database for that entity.
And the data for each field is loaded, ready to be modifed.
- the `new` button load the same form as explain just above, but virgin.
- the `new` form and `modify` form has a `save changes` button who
load `controller_dashboard_form.php` who will save the datas for each
field in the database table of the entity (in a new line or in the
modified line, according to situation).

NEXT STEP :
- the field `active` could use a dropdown menu.
- the field `fk_group_rowid` and similar foreign keys could use a
dropdown menu as well.