(page_bio) add many var_dump to find th 'job' bug.
When loading the page 'bio' , the first
instance created (for Groupe Alinea) was
showing a bug (exception) cause the
that field had no value. I first modify the
class Bio to render an empty string,
but same bug.
Finally, the reason was that in class Bio,
I forgot to put the property for 'job'
(even if I did add 'job' in all the many
other places.
Anyway, now it's fixed.